+++ pkg/convertor/words/languages/languages.go
@@ -1,0 +1,7 @@
+package languages
+
+type Language string
+
+const (
+	Russia = Language(`ru`)
+)