+++ pkg/convertor/words/ordinal_numbers.go
@@ -1,0 +1,18 @@
+package words
+
+import (
+	"github.com/ThCompiler/go_game_constractor/pkg/convertor/constants"
+	"github.com/ThCompiler/go_game_constractor/pkg/convertor/words/declension"
+	"github.com/ThCompiler/go_game_constractor/pkg/convertor/words/genders"
+)
+
+type declensionOrdinalNumbersT map[declension.Declension][constants.CountScaleNumberNameForms]string
+
+type genderOrdinalNumbersT map[genders.Gender]declensionOrdinalNumbersT
+
+type OrdinalNumbers struct {
+	Units    []genderOrdinalNumbersT `yaml:"units,omitempty"`
+	Tens     []genderOrdinalNumbersT `yaml:"tens,omitempty"`
+	Dozens   []genderOrdinalNumbersT `yaml:"dozens,omitempty"`
+	Hundreds []genderOrdinalNumbersT `yaml:"hundreds,omitempty"`
+}