+++ scg/cmd/main.go
@@ -12,9 +12,10 @@ import (
 
 func main() {
 	var (
-		out     = flag.String("output", "", "")
-		script  = flag.String("script", "", "")
-		version = flag.Bool("version", true, "")
+		out     = flag.String("output", "", "Path to directory, where you want generate code for script")
+		script  = flag.String("script", "", "Path to config file as yaml, json or xml with info about script")
+		update  = flag.Bool("update", false, "Return current version")
+		version = flag.Bool("version", false, "Return current version")
 	)
 	{
 		flag.Parse()
@@ -37,7 +38,7 @@ func main() {
 		fail(err.Error())
 	}
 
-	outputs, err := generator.Generate(*out, *si)
+	outputs, err := generator.Generate(*out, *si, *update)
 	if err != nil {
 		fail(err.Error())
 	}